Make orca build use the gnome-44 branch
After cloning Orca, switch to the gnome-44 branch. This branch is a backward compatibility version of Orca for systems with older at-spi libraries and other dependencies, so LTS systems like Ubuntu Mate 22.04 could still receive bugfixes for the screenreader. UM22.04 doesn't use the latest accessibility stack, therefore this branch is necessary for the build to work correctly.