Fix Orca autostart on UM 24.04
Ubuntu Mate 24.04 suffers from a bug which disables the org.gnome.desktop.a11y.applications screen-reader-enabled key upon the system startup. This key is toggled by the Super+Alt+S shortcut and controls whether the screenreader should be autostarted or not, depending on the user configuration. However, since the value seems to be reset even if explicitly set by providing a system gscheme override, this commit fixes the issue by introducing a modified orca autostart ignoring the variable. The gsetting is additionally set as well upon the system startup to ensure the proper functionality of the Super+Alt+S shortcut.